requirements/testing.txt
-r base.txt

django-extensions==2.2.1
Werkzeug==0.15.6
${IA2_TEST_MODEL_FILE}
